feature: splitting the app into components
