+++ Investment_App/src/components/Header/Header.jsx
@@ -1,0 +1,12 @@
+import logo from "../../assets/investment-calculator-logo.png";
+
+const Header = () => {
+  return (
+    <header className="header">
+      <img src={logo} alt="logo" />
+      <h1>Investment Calculator</h1>
+    </header>
+  );
+};
+
+export default Header;
